refactor(reviews): rename outbound/kafka -> outbound/messaging
The adapter is no longer Kafka-specific — it now holds a
Publisher interface that may resolve to kafka or jetstream.
