+++ services/reviews/internal/adapter/inbound/http/handler_test.go
@@ -12,7 +12,7 @@ import (
 
 	"github.com/kaio6fellipe/event-driven-bookinfo/pkg/idempotency"
 	handler "github.com/kaio6fellipe/event-driven-bookinfo/services/reviews/internal/adapter/inbound/http"
-	"github.com/kaio6fellipe/event-driven-bookinfo/services/reviews/internal/adapter/outbound/kafka"
+	"github.com/kaio6fellipe/event-driven-bookinfo/services/reviews/internal/adapter/outbound/messaging"
 	"github.com/kaio6fellipe/event-driven-bookinfo/services/reviews/internal/adapter/outbound/memory"
 	"github.com/kaio6fellipe/event-driven-bookinfo/services/reviews/internal/core/domain"
 	"github.com/kaio6fellipe/event-driven-bookinfo/services/reviews/internal/core/service"
@@ -39,7 +39,7 @@ func setupHandler(t *testing.T) *http.ServeMux {
 			},
 		},
 	}
-	svc := service.NewReviewService(repo, client, idempotency.NewMemoryStore(), kafka.NewNoopPublisher())
+	svc := service.NewReviewService(repo, client, idempotency.NewMemoryStore(), messaging.NewNoopPublisher())
 	mux := http.NewServeMux()
 	h := handler.NewHandler(svc)
 	h.RegisterRoutes(mux)
